[Fix] update imports after LGCA refactor
